rewrited RGFBinaryClassifier class hierarchy
